requirements.txt
# ---------- Core ----------
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
gunicorn>=22.0,<24.0
pydantic>=2.0,<3.0
websockets>=12.0,<14.0

# ---------- Database ----------
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0

# ---------- Data ----------
pandas>=2.2,<3.0
numpy>=1.26,<3.0
yfinance>=0.2,<1.0
scipy>=1.12,<2.0

# ---------- ML ----------
lightgbm>=4.3,<5.0
xgboost>=2.0,<3.0
scikit-learn>=1.4,<2.0
optuna>=3.6,<4.0
shap>=0.45,<1.0
torch>=2.2,<3.0

# ---------- Monitoring ----------
prometheus-client>=0.20,<1.0
mlflow>=2.12,<3.0
sentry-sdk[fastapi]>=2.0,<3.0

# ---------- Task Queue ----------
celery[redis]>=5.3,<6.0

# ---------- HTTP ----------
requests>=2.31,<3.0
httpx>=0.27,<1.0

# ---------- Broker ----------
smartapi-python>=1.3,<2.0
pyotp>=2.9,<3.0
logzero>=1.7
websocket-client>=1.6
python-dotenv>=1.0,<2.0

# ---------- Testing ----------
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0

# ---------- Linting ----------
flake8>=7.0,<8.0
